completed the bar graph
