+++ app/component/Layout.js
@@ -63,8 +63,8 @@ export default function Layout({ children }) {
     { label: 'Users', path: '/user', icon: <PersonIcon /> },
     { label: 'Post', path: '/post', icon: <DataIcon /> },
     { label: "Table", path: "/table", icon: <TableChartIcon /> },
-    { label: "Chart", path: "/chart", icon: <TableChartIcon /> },
-    { label: "BarChart", path: "/barchart", icon: <DashboardIcon /> },
+    { label: "Line Chart", path: "/linegraph", icon: <TableChartIcon /> },
+    { label: "Bar Chart", path: "/bargraph", icon: <DashboardIcon /> },
   ];
 
   return (
