+++ app/constant/bardata.json
@@ -2933,7 +2933,7 @@
         "deviceId": "6577e92789dd96f6e35d90c6",
         "testDateTime": "2025-09-01T07:33:40.507Z",
         "readingFromDevice": 26.75,
-        "calibratedReading": 26.75,
+        "calibratedReading": 25.80,
         "calibrationPercentage": 0,
         "isVisible": true,
         "measuringUnitFullName": "Celsius",
@@ -2958,7 +2958,7 @@
         "deviceId": "6577e92789dd96f6e35d90c6",
         "testDateTime": "2025-09-01T07:29:02.202Z",
         "readingFromDevice": 36,
-        "calibratedReading": 45,
+        "calibratedReading": 55,
         "calibrationPercentage": 0,
         "isVisible": true,
         "measuringUnitFullName": "Celsius",
